Update patch set 1
Patch Set 1:

(2 comments)

So we'll need this to get a newer version of qemu for Mitaka users. We had it already on RDO trunk.

A couple unrelated commits, we could disregard them and treat them as a follow-up, since this is a cherry-pick.

Patch-set: 1
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0