Add support for c8-stream
A post install script is added to detect distro
and set a dnf var named 'cloudsigdist' which is used
to set up repos based on it.
For centos stream and non centos distro use '8-stream'
else '8'.

Remove dnf var as part of post uninstall.

Also fix release in commented baseurl.

Change-Id: I4cb6de95f4215dd5bc429cfa47c9f6a9225f9233