Add support for c8-stream
A post install script is added to detect distro
and set a dnf/yum var named 'cloudsigdist' which is used
to set up repos based on it.
For centos stream and non centos distro use '8-stream',
for el7 use '7' else '8'.

Remove dnf/yum vars as part of post uninstall.

As we are not pushing content to centos8-stream
mirrors for openstack-train as it's about to EOL
in next few months, not changing release repo, only
testing repo will use 'cloudsigdist' yum var.

Change-Id: I097e2a26902420519eeaf82c64ef54af2a7a9dcc